+++ docs/conf.py
@@ -1,7 +1,7 @@
 project = "DBHose-Airflow"
 copyright = "2025-%Y, 0xMihalich"
 author = "0xMihalich"
-release = "0.2.0.dev4"
+release = "0.2.0.dev5"
 master_doc = "index"
 language = "ru"
 html_theme = "alabaster"