+++ src/dbhose_airflow/core/__init__.py
@@ -7,6 +7,16 @@ from .ddl_core import (
     postgres_ddl,
     postgres_sequence_ddl,
 )
+from .move import (
+    AppendStrategy,
+    AutoStrategy,
+    CustomStrategy,
+    DeleteStrategy,
+    ReplaceStrategy,
+    RewriteStrategy,
+    MoveStrategy,
+    get_move_strategy,
+)
 from .structs import (
     ConnectionConfig,
     ColumnMeta,
@@ -22,19 +32,27 @@ from .structs import (
 
 
 __all__ = (
+    "AppendStrategy",
+    "AutoStrategy",
     "ConnectionConfig",
     "ColumnMeta",
+    "CustomStrategy",
+    "DeleteStrategy",
     "DQCheck",
     "DQConfig",
     "DQTest",
     "Error",
     "ETLInfo",
+    "MoveStrategy",
     "MoveMethod",
     "MoveType",
+    "ReplaceStrategy",
+    "RewriteStrategy",
     "StagingConfig",
     "TableMetadata",
     "clickhouse_ddl",
     "generate_ddl",
+    "get_move_strategy",
     "postgres_ddl",
     "postgres_sequence_ddl",
 )